+++ software/shared/src/lib.rs
@@ -20,7 +20,7 @@ pub struct DashboardData {
     pub mem_used: u16,
     pub swap_used: u16,
     pub cpu_load: u16,
-    pub cpu_temp: u8,
+    pub cpu_temp: u16,
     pub player_status: String<8>,
     pub title: String<64>,
     pub artist: String<64>,
